Style hero section
Styled hero section. Moved the font declaration to body element since
the same font is used across the page anyway. Added a general Call-To-Action
color variable. Put a padding-bottom system in place for positioning
sections of the website because of the many background colors used around
the page. This helps avoid a situation where margin-bottom leaves a gap
filled with body's background color between sections.